Extract chathubservice from chathub
